XXX wrong about 5.25: Initialize locale object even in unthreaded perls
... and even if not all the thread-safe components are available.

5.25, in order to avoid race conditions.

This commit will now initialize the thread-safe C locale object if the
function to do so is available, instead of only if all functions for any
manipulation are needed.  This will allow for a future commit that needs
only a subset of them, and which is a win on unthreaded builds.