locale.c: Refactor some #if clauses
This moves all the handling of the case where there are no locale
messages, instead of splitting it up across long stretches of
conditionally compiled code.  This code is essentially trivial, and seen
to be when it isn't split up, and this prepares for the next commit.

The final return of the function is still split off so that all branches
go through it, and the debugging code adjacent to it.