locale.c: Don't be too clever in strlcat
This code I wrote was attempting to avoid multiple calls to strlen in
constructing the catenation of various components of a string.  It did
this by keeping track of how far it got each iteration, and using that
as a starting point for the next.  I now realize that the return value
of strlcat is as if it succeeds, even if there isn't enough room.  That
means that if there were a problem, this could start out an iteration
such that it would be writing beyond the end of the buffer.  It is safer
to not do this, so this commit removes it.

The use of strlcat is a safety measure, as there should be a sufficient
amount of space calculated for things to fit, so there is no bug here.
But one should be safe.